pyproject.toml
[tool.poetry]
name = "olistflix_users"
version = "0.1.0"
description = ""
authors = ["Lucas Dario"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.6"
djangorestframework = "^3.12.2"
autopep8 = "^1.5.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
